Transfer of pagination from movieService and movieList.component to pagination.component. Fixed movieList component
