+++ src/app/movie-list/movie-list.component.ts
@@ -2,6 +2,7 @@ import {Component, OnInit} from '@angular/core';
 import {Observable} from 'rxjs/Observable';
 import {MovieService} from '../services/movie.service';
 import {Movie} from '../services/movie';
+// import * as _ from 'underscore';
 
 
 @Component({
@@ -16,20 +17,19 @@ export class MovieListComponent implements OnInit {
   total_pages: number;
   page: number;
   total_results: number;
-  pager: any = {};
+  searchValue: string;
+  // pager: any = {};
 
-  constructor(
-    private movieService: MovieService
-  ) {
+  constructor(private movieService: MovieService) {
   }
 
   ngOnInit() {
     this.getMovies();
-    this.setPage(1);
+    // this.setPage(this.page, this.total_pages);
   }
 
-  getMovies(searchValue?: string, page: number = 1) {
-    this.movieService.getMovies(searchValue)
+  public getMovies() {
+    this.movieService.getMovies(this.searchValue, this.page)
       .subscribe(
         response => {
           this.movieList = response['results'];
@@ -39,11 +39,53 @@ export class MovieListComponent implements OnInit {
         });
   }
 
-  setPage(page: number) {
-    if (page < 1 || page > this.pager.totalPages) {
-      return;
-    }
-    // get pager object from service
-    this.pager = this.movieService.getPager(this.page, this.total_pages);
+  public setPage(page: number){
+    this.page = page;
+    this.getMovies();
   }
+
+  public setSearchValue(searchValue: string){
+    this.page = 1;
+    this.searchValue = searchValue;
+    this.getMovies();
+  }
+
+  // setPage(page: number, total_pages: number) {
+  //   if (page < 1 || page > this.pager.totalPages) {
+  //     return;
+  //   }
+  //   this.pager = this.getPager(page, total_pages);
+  //   this.getMovies();
+  //     console.log(this.pager);
+  // }
+  // getPager(page: number = 1, total_pages: number){
+  //   let currentPage = page;
+  //   let totalPages = total_pages;
+  //
+  //   let startPage: number, endPage: number;
+  //
+  //   if (totalPages <= 5) {
+  //     startPage = 1;
+  //     endPage = totalPages;
+  //   } else {
+  //     if (currentPage <= 3) {
+  //       startPage = 1;
+  //       endPage = 5;
+  //     } else if (currentPage + 1 >= totalPages) {
+  //       startPage = totalPages - 4;
+  //       endPage = totalPages;
+  //     } else {
+  //       startPage = currentPage - 2;
+  //       endPage = currentPage + 2;
+  //     }
+  //   }
+  //   let pages = _.range(startPage, endPage + 1);
+  //   return {
+  //     currentPage: currentPage,
+  //     totalPages: totalPages,
+  //     startPage: startPage,
+  //     endPage: endPage,
+  //     pages: pages
+  //   };
+  // }
 }
