+++ src/app/pagination/pagination.component.css
@@ -1,3 +1,6 @@
 .pagination li.active {
   background-color: #20B2AA;
 }
+a {
+  cursor: pointer;
+}
