+++ src/app/services/movie.service.ts
@@ -3,7 +3,7 @@ import {Http, Response} from '@angular/http';
 import 'rxjs/add/operator/map';
 import {Observable} from 'rxjs/Observable';
 import {Movie} from './movie';
-import * as _ from 'underscore';
+
 
 @Injectable()
 export class MovieService {
@@ -26,43 +26,5 @@ export class MovieService {
   private getDataByUrl(url: string): Observable<Movie[]>{
     return this.http.get(url)
       .map((res) => { return res.json() });
-    // .map(this.extractData);
   }
-
-  public getPager(page: number = 1, total_pages: number) {
-    // calculate total pages
-    let currentPage = page;
-    let totalPages = total_pages;
-
-    let startPage: number, endPage: number;
-
-    if (totalPages <= 5) {
-      startPage = 1;
-      endPage = totalPages;
-    } else {
-      if (currentPage <= 3) {
-        startPage = 1;
-        endPage = 5;
-      } else if (currentPage + 1 >= totalPages) {
-        startPage = totalPages - 4;
-        endPage = totalPages;
-      } else {
-        startPage = currentPage - 2;
-        endPage = currentPage + 2;
-      }
-    }
-    let pages = _.range(startPage, endPage + 1);
-    return {
-      currentPage: currentPage,
-      totalPages: totalPages,
-      startPage: startPage,
-      endPage: endPage,
-      pages: pages
-    };
-  }
-
-  // private extractData(res: Response) {
-  //   let body = res.json();
-  //     return body.results;
-  // }
 }
